Update dependencies and rebuild container w/ new PUDL
Merged a bunch of changes into jupyterhub-beta branch in the PUDL repo,
so this will update the container build. Also bumped some dependencies
to be in line with what's now being used on sprint30. Will need to
upload new versions of the databases and parquet to be fully compatible,
which might take a few hours.

Also removed a line from the 2nd notebook that was printing out the
user's EIA API key, just to avoid accidentally publicizing that with a
check-in.