Expose notebook dir in container and remove port from .env
Since we're setting a particular (random) port in the docker-compose.yml
I removed the option to set it with a user selected environment variable
in the .env file.

Added a bind mount to expose the notebooks directory within the
container, so that the user doesn't have to think about what's inside
vs. outside of the container -- it should pretty much entirely be
referencing file which are in their filesystem.

Also forced jupyterlab to use at least v3.0.12 in the environment.yml
file because for some reason it only installed 3.0.7... not sure why?