Add dask-gateway to image
This image is used by the catalyst coop hub used by 2i2c, and we are now starting
to automatically test the hubs on each change. As dask-gateway functionality is
available for this hub, we test it too - but as the `dask-gateway` package isn't
installed in the image, it fails! This PR adds `dask-gateway` to the image, and
will help keep our infra more stable.